Refactor: Move application code to app folder
