+++ app/models/crud.py
@@ -1,11 +1,12 @@
 import uuid
 import datetime
 from functools import lru_cache
+
 from sqlalchemy.orm import Session
 from sqlalchemy import exc, func
-from . import models, schemas
-from .database import HashableSession
-from utils import add_ratings, get_movie_download
+
+from app.models import models, schemas, HashableSession
+from app.utils import add_ratings, get_movie_download
 
 
 def get_movie(db: Session, movie_id: int):
@@ -20,25 +21,23 @@ def list_movies(db: HashableSession, engine: str, page: int, num: int):
     """
     Get List of Movies. With pagination and a total of `num` rows per query
     """
-    return list(db.query(models.Movie)\
-                .filter(
-                    func.lower(models.Movie.engine) == engine.lower())\
+    return list(db.query(models.Movie)
+                .filter(func.lower(models.Movie.engine) == engine.lower())
                 .limit(num).offset(num * (page-1)))
 
+
 @lru_cache(maxsize=1000)
-def search_movies(db: HashableSession, engine: str, query: str, page:int, num: int):
+def search_movies(db: HashableSession, engine: str, query: str, page: int, num: int):
     """
     Search movies using fuzzy partial
     """
-    ret =  list(db.query(models.Movie)\
-        .filter(\
-            models.Movie.name.ilike("%"+query+"%"),
-            func.lower(models.Movie.engine) == engine.lower()
-            )\
-        .limit(num)\
-        .offset(num *(page-1))\
-        )
-    return ret
+    # TODO Improve fuzzy searching
+    return list(db.query(models.Movie)
+                .filter(models.Movie.name.ilike("%"+query+"%"), func.lower(models.Movie.engine) == engine.lower())
+                .limit(num)
+                .offset(num * (page-1))
+                )
+
 
 @lru_cache(maxsize=200)
 def get_movie_by_referral_id(db: HashableSession, referral_id: str):
