+++ app/models/schemas.py
@@ -3,17 +3,21 @@ from typing import List, Optional
 
 from pydantic import BaseModel
 
-### Rating schemas
+# Rating schemas
+
 
 class RatingBase(BaseModel):
     ip_address: str
+
 
 class SpecificRating(RatingBase):
     movie_name: str
     engine: str
 
+
 class IndexedRating(RatingBase):
     movie_id: int
+
 
 class SpecificRatingScore(SpecificRating):
     score: str
@@ -23,6 +27,7 @@ class SpecificRatingScore(SpecificRating):
     download_link: str
     cover_photo_link: str
 
+
 class Rating(RatingBase):
     id: int
     movie_id: int
@@ -31,9 +36,11 @@ class Rating(RatingBase):
     class Config:
         orm_mode = True
 
+
 class AverageRating(BaseModel):
     average_ratings: float
     by: int
+
 
 class RatingCreate(RatingBase):
     movie_name: str
@@ -41,10 +48,11 @@ class RatingCreate(RatingBase):
     score: int
 
 
-##### Download Schemas
+# Download Schemas
 
 class DownloadBase(BaseModel):
     ip_address: str
+
 
 class DownloadCreate(DownloadBase):
     movie_name: str
@@ -54,6 +62,7 @@ class DownloadCreate(DownloadBase):
     year: str
     download_link: str
     cover_photo_link: str
+
 
 class DownloadFilter(BaseModel):
     filter_by: str
@@ -66,10 +75,11 @@ class DownloadFilter(BaseModel):
 
     def __hash__(self):
         return hash(repr(self))
-    
+
     def __eq__(self, other):
         print(self.__hash__(), other.__hash__())
         return self.__hash__() == other.__hash__()
+
 
 class Download(DownloadBase):
     id: int
@@ -79,10 +89,12 @@ class Download(DownloadBase):
     class Config:
         orm_mode = True
 
-###### Referral Schemas
+# Referral Schemas
+
 
 class ReferralBase(BaseModel):
     ip_address: str
+
 
 class ReferralCreate(ReferralBase):
     movie_name: str
@@ -93,6 +105,7 @@ class ReferralCreate(ReferralBase):
     download_link: str
     cover_photo_link: str
 
+
 class Referral(ReferralBase):
     id: int
     movie_id: int
@@ -102,11 +115,12 @@ class Referral(ReferralBase):
         orm_mode = True
 
 
-##### Movie Schemas
+# Movie Schemas
 
 class MovieBase(BaseModel):
     name: str
     engine: str
+
 
 class MovieCreate(MovieBase):
     description: Optional[str]
@@ -115,15 +129,17 @@ class MovieCreate(MovieBase):
     download_link: Optional[str]
     cover_photo_link: Optional[str]
 
+
 class MovieComplete(MovieCreate):
     id: int
     referral_id: Optional[str]
-    
+
 
 class MovieReferral(MovieComplete):
 
     class Config:
         orm_mode = True
+
 
 class MovieDownloads(MovieComplete):
     downloads: int
